backend/requirements.txt
# Web Framework
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
starlette==0.26.1

# Database
sqlalchemy==2.0.9
alembic==1.10.3
psycopg2-binary==2.9.6
pymongo==4.3.3

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# ML Integrations
mlflow==2.3.1
scikit-learn==1.2.2
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1

# Utilities
python-dotenv==1.0.0
requests==2.28.2
loguru==0.7.0
pyyaml==6.0
pytest==7.3.1
pytest-cov==4.1.0 